Make repo troubleshooting a bit easier
While load testing the repo subcommands it was discovered
that reprepro does create a log file to trouble shoot problems.
This makes the operator experience difficult to
use since it's not easy to figure out what's going on.
Do not hide the output of the reprepo when running
through the CLI.

Also fix some CLI issues when displaying the list of
packages that are apart of a repository.

Story: 2010867
Task: 48556

Depends-On: https://review.opendev.org/c/starlingx/apt-ostree/+/894741

Change-Id: Ifb34ee39e5197e24efe63537f27b2b480d46ed86
Signed-off-by: Charles Short <charles.short@windriver.com>