Add "export/import" support
Add the ability to export and import ostree branches into
an ostree repository.

The way that this works is that apt-ostree uses "the ostree
pull-local" command to import or export an ostree branch
into a repository. This allows the user to easily import a branch
from an extermal source, like an ISO.

The methods that are called "backup" and "restore" so the CLI
library wont get traceback when trying to run them.

Story: 2010867
Task: 48556

Change-Id: Ic1791089118a64ab491673abcaec1559c725eaa9
Signed-off-by: Charles Short <charles.short@windriver.com>